Merge approved after a conversation with PM Hank earlier today, 6/25/19.

+++ api/routes/contractorRoutes.js
@@ -6,7 +6,15 @@ const router = express.Router();
 router.get('/', async (req, res) => {
   try {
     const contractors = await query('SELECT * FROM contractors;');
-    return res.json({ contractors: contractors.rows });
+    const promises = contractors.rows.map(async contractor => {
+      const services = await query(
+        'SELECT * FROM services WHERE "contractorId" = $1',
+        [contractor.id]
+      );
+      return { ...contractor, services: services.rows };
+    });
+    const contractorsWithServices = await Promise.all(promises);
+    return res.json({ contractors: contractorsWithServices });
   } catch (error) {
     return res.status(500).json({ error: error.message });
   }
@@ -33,10 +41,16 @@ router.get('/:id', async (req, res) => {
     const contractor = await query('SELECT * FROM contractors WHERE id = $1;', [
       id,
     ]);
-    if (!contractor.rows || !contractor.rows.length) {
+    if (!contractor.rows || !contractor.rows[0]) {
       throw new Error(404);
     }
-    return res.json({ contractor: contractor.rows });
+    const services = await query(
+      'SELECT * FROM services WHERE "contractorId" = $1',
+      [contractor.rows[0].id]
+    );
+    return res.json({
+      contractor: { ...contractor.rows[0], services: services.rows },
+    });
   } catch (error) {
     switch (error.message) {
       case '404':
@@ -59,7 +73,6 @@ router.post('/', async (req, res) => {
       stateAbbr,
       zipCode,
     } = req.body;
-    // const { id } = req.decoded;
     if (
       !contractorName ||
       !phoneNumber ||
@@ -73,10 +86,6 @@ router.post('/', async (req, res) => {
       'INSERT INTO contractors (name, "phoneNumber", "streetAddress", city, "stateAbbr", "zipCode") VALUES ($1, $2, $3, $4, $5, $6) RETURNING *',
       [contractorName, phoneNumber, streetAddress, city, stateAbbr, zipCode]
     );
-    // await query(`UPDATE users SET contractorId = $1 WHERE id = $2;`, [
-    //   contractor.rows[0].id,
-    //   id,
-    // ]);
     return res.status(201).json(contractor.rows[0]);
   } catch (err) {
     switch (err.message) {
@@ -95,6 +104,7 @@ router.post('/', async (req, res) => {
 
 router.put('/:id', async (req, res) => {
   const { id } = req.params;
+  if (id !== req.user.contractorId) throw new Error(403);
   try {
     const contractor = await query(
       'UPDATE contractors SET name = $1, "phoneNumber" = $2, "streetAddress" = $3, city = $4, "stateAbbr" = $5, "zipCode" = $6 WHERE id= $7 RETURNING *',
@@ -111,6 +121,8 @@ router.put('/:id', async (req, res) => {
     return res.json({ contractor: contractor.rows[0] });
   } catch (error) {
     switch (error.message) {
+      case '403':
+        return res.status(403).json({ error: 'Forbidden' });
       case '404':
         return res
           .status(404)
@@ -121,48 +133,10 @@ router.put('/:id', async (req, res) => {
   }
 });
 
-// Note: Use this one if the one above doesnt work
-// router.put('/:id', async (req, res) => {
-//   try {
-//     const contractor = await query(
-//       'UPDATE contractors SET name = ($1), phoneNumber = ($2), streetAddress = ($3), city = ($4), stateAbbr = ($5), zipCode = ($6) WHERE id = ($7) RETURNING *',
-//       [
-//         req.body.name,
-//         req.body.phoneNumber,
-//         req.body.streetAddress,
-//         req.body.city,
-//         req.body.stateAbbr,
-//         req.body.zipCode,
-//         req.params.id,
-//       ]
-//     );
-//     return res.json(contractor.rows[0]);
-//   } catch (err) {
-//     return err;
-//   }
-// });
-
-// As a callback. Note:  save for back up endpoint
-// router.put('/:id', (req, res) => {
-//   const id = parseInt(req.params.id);
-//   const {name, phoneNumber, streetAddress, city, stateAbbr, zipCode} = request.body};
-
-//   query(
-//   'UPDATE contractors SET name = $1, phoneNumber = $2, streetAddress = $3, city = $4, stateAbbr = $5, zipCode = $6 WHERE id = $7',
-//   [name, phoneNumber, streetAddress, city, stateAbbr, zipCode],
-//   (error, result) => {
-//     if (error) {
-//       throw error
-//     }
-//     res.status(200).send(`contractor modified with ID: ${id}`)
-//   }
-//   )
-// });
-
-// As a callback
 router.delete('/:id', async (req, res) => {
   try {
     const { id } = req.params;
+    if (req.user.contractorId !== id) throw new Error(403);
     const contractor = await query(
       'DELETE FROM contractors WHERE id = $1 RETURNING *',
       [id]
@@ -171,6 +145,8 @@ router.delete('/:id', async (req, res) => {
     return res.json({ deleted: contractor.rows[0] });
   } catch (err) {
     switch (err.message) {
+      case '403':
+        return res.status(403).json({ error: 'Forbidden' });
       case '404':
         return res
           .status(404)
